fix kernel paging
The paging was broken even though the the tlb from qemu shows it has the
correct memory mapping. The issue is caused by using the fixed vga
buffer address 0xb8000 which is not mapped by the page tale, so after we
switched the new page table the tty module writes to unmapped memory.

The fix is to set the vga buffer address to its virtual address. It is
good practice to always stay in vritual address in kernel mode. because
this kind of paging bug is very hard to debug.