add init1
Initialize the first process init1.

Note: The physical address allocator seems to be broken, it can only
allocate up to 0x88000000, if you allocate pages over the addess the
freelist's first next page seems to always point to 0.

in qemu, x/1b 0x88000000 gives a memory not accessible error, but after
checking the tlb it seems the virtual address is mapped in the page table.

I modified the `PHYSTOP` from 0xE000000 to 0x8000000, so the kernel virutal
ddress now ends at 0x88000000. But it is an temporary fix, the allocator
needs to be fixed to move on.

PS: This PR also fixed some bugs among utility functions:
        1. string functions in "string.h" uses `uint32_t` for size cause
           pattern like `while (n--)` to loop forever. We changed it to
           signed integer.
        2. `tty_printf` handles `%s` case incorrectly. Instead of
           increment the char * pointer, it sets the pointer to the end
           of the formatted string. We corrected this behavior.