Fix vmfree bug
It's because when we deallocate user vm, we deallocate from KERN_BASE
all the way to 0, so some pages are outside of the possible phsyical
memory range (end - PHSYTOP). Even though the virtual memory address
space is 0 to KERN_BASE, we can't address the entire range because
of the PHSYTOP limitation.